backend/requirements.txt
# FastAPI and supporting tools
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2

# MongoDB Connection
pymongo==4.6.0

# AI/LLM - Google's Generative AI
google-genai==0.6.0

# Utilities
pillow==10.1.0
python-dotenv==1.0.0

# Vector Embeddings and RAG Libraries - Using the latest versions available
# Core AI Dependencies (Pinning to KNOWN STABLE COMBO)
numpy==1.26.4
transformers==4.35.2      # <--- Stable Version
sentence-transformers==2.2.2 # <--- Stable Version
huggingface-hub==0.17.3   # <--- Stable Version for older s-t
torch==2.3.0              # <--- A recent, well-tested version
faiss-cpu==1.11.0         # <--- Version confirmed available